src/core/kindwise_wrapper.py
from kindwise.plant import PlantApi
from src.config import settings

class KindwiseClient:
    def __init__(self):
        """
        Initialize the Kindwise API client.
        
        TODO:
        - Configure the Kindwise API client with the API key.
        """
        # TODO: Initialize PlantApi with the API key from settings
        self.api = None  # Placeholder for Kindwise API client

    def identify_plant(self, image_path: str, lat_lon: tuple = None):
        """
        Identify the plant from the given image.

        TODO:
        - Send image to Kindwise API for plant identificroptation.
        - Handle errors or retries in case of failures.
        
        Placeholder:
        Returns mock identification data.
        """
        # TODO: Call Kindwise API to identify plant
        return {
            "plant_name": "Placeholder Plant",
            "scientific_name": "Plantae Placeholder",
            "accuracy": 0.95
        }
