requirements.txt
fastapi
uvicorn
pymongo
pydantic
pillow
python-dotenv
pydantic-settings
requests
python-multipart 
